+++ components/Header/setup.ts
@@ -4,6 +4,7 @@ import { LAYOUT_TYPES } from '../../src/enums.ts';
 
 export type iHeader = iExtendedElement & {
   layout_type: LAYOUT_TYPES | null;
+  banner: boolean;
   fwd: Partial<{
     layout: iFwd<HTMLDivElement>;
     panel: iFwd<HTMLDivElement>;
@@ -12,6 +13,7 @@ export type iHeader = iExtendedElement & {
 
 const defaults: iHeader = {
   layout_type: null,
+  banner: false,
   fwd: {},
 };
 
@@ -20,8 +22,8 @@ export default (props: Partial<iHeader>) => {
 
   const classes = {
     header: cn('comp-header', p.class),
-    layout: cn('', p.fwd.layout?.class),
-    panel: cn('', p.fwd.panel?.class),
+    layout: cn(p.fwd.layout?.class),
+    panel: cn(p.banner ? 'comp-header_banner' : null, p.fwd.panel?.class),
   };
 
   return { c: classes, ...p };